Raspberry Pi Samsung TV IR Remote Control.
This is incomplete Proof Of Concept Code, designed to prove that you
can manually manipulate Raspberry Pi GPIO pins straight from Python.

The code replicates Samsung IR Remote Control signals using an infra-red
LED connected to a resistor and GPIO Pins.

NOTE: this does not use a physical PWM generator to generate the sub-carrier
IR frequency.  It's done using Rasbian, a Raspberry Pi 3B and a few techniques
to 'encourage' the system to generate the correctly timed pulses.

This project was created by putting an oscilloscope onto the IR LED output
of a Samsung TV remote control (2014 vintage, or so...).  After the IR
packet form and sub-carrier frequency was read, the word formation
and individual button digital sequence was manually read.

The Python code is more designed to be 'verbose' - namely very manual,
less automated and easy to read.
- That said, it was done for myself and not for publishing.  It is here
because I mentioned it in a redit post and I was asked to share the code.
So here it is...

Happy for people to improve the code - but please only do so if you retain
the verbose nature of variable names, etc.  Readability is essential to
ensure the method is easy to understand.

Also, please share any manual IR remote control button / bit reads, and
(if demand exists), we could have some fun and create a smart home legacy
device control system.

Incidently, I linked this to Python's web app, Flask, and https://ifttt.com/
which links nicely to Google Nest Hub (etc), which meant I could control
my TV via voice....

Anyway, its here and if of interest, I'd be happy to participate in
growing it further. Or just enjoy the practical outcome of this rather
simple code..